refactor: updated error resolver package name
