+++ src/main/java/com/nephren/raven/apiclient/errorresolver/ApiErrorResolver.java
@@ -1,4 +1,4 @@
-package com.nephren.raven.apiclient.error;
+package com.nephren.raven.apiclient.errorresolver;
 
 import java.lang.reflect.Method;
 import reactor.core.publisher.Mono;
