+++ src/test/java/com/nephren/raven/apiclient/serviceExample/client/errorresolver/DeleteExampleClientErrorResolver.java
@@ -1,6 +1,6 @@
 package com.nephren.raven.apiclient.serviceExample.client.errorresolver;
 
-import com.nephren.raven.apiclient.error.ApiErrorResolver;
+import com.nephren.raven.apiclient.errorresolver.ApiErrorResolver;
 import java.lang.reflect.Method;
 import org.springframework.stereotype.Component;
 import reactor.core.publisher.Mono;
